Update to sitespeed.io 4.0.0-alpha5
The 3.x version of sitespeed was depending on a version of slimerjs that is affected by https://github.com/graingert/slimerjs/issues/39#issuecomment-224910698.

After I realized that sitespeed has been in alpha for a while and that the new version is not affected I decided to try bumping and this seems to have fixed the build. I'm not sure about the downstream changes needed for the new sitespeed.io package though.